+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -139,7 +139,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-23T05:38:05.558Z",
+    "2026-01-23T09:49:01.865Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {