+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -201,7 +201,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfmnrkkfmer4v6ey2mkxhwhb"),
+  _id: z.string().default("conv_01kfn442kxpdt9f247v9zqe0az"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),