+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-01-23T05:38:09.460Z",
+    "2026-01-23T09:49:05.510Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {