+++ packages/orq-rc/src/models/operations/listcontacts.ts
@@ -226,7 +226,7 @@ export const Data$inboundSchema: z.ZodType<Data, z.ZodTypeDef, unknown> = z
     created: z.string().datetime({ offset: true }).transform(v => new Date(v))
       .optional(),
     updated: z.string().datetime({ offset: true }).default(
-      "2026-01-23T05:38:05.558Z",
+      "2026-01-23T09:49:01.865Z",
     ).transform(v => new Date(v)),
     metrics: z.lazy(() => ListContactsMetrics$inboundSchema),
   }).transform((v) => {