+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4015,7 +4015,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KFMNRKE3ZJWC8EQYXDJC06T7"),
+  id: z.string().default("01KFN442EXBERB0Z8MPCVT0CS1"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),