+++ packages/orq-rc/RELEASES.md
@@ -6149,3 +6149,13 @@ Based on:
 - [typescript v4.2.0-rc.64] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.64] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.64 - packages/orq-rc
+
+## 2026-01-23 09:51:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.696.0 (2.799.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.65] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.65] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.65 - packages/orq-rc